Revert "Update: Move pattern actions to the editor package. (#60785)"
This reverts commit 66ade7093fb80dd656169c3ded366bd2e7ec3cd5.
