+++ packages/edit-site/package.json
@@ -69,6 +69,7 @@
 		"@wordpress/widgets": "file:../widgets",
 		"@wordpress/wordcount": "file:../wordcount",
 		"change-case": "^4.1.2",
+		"client-zip": "^2.4.4",
 		"clsx": "^2.1.1",
 		"colord": "^2.9.2",
 		"fast-deep-equal": "^3.1.3",
