+++ packages/edit-site/src/components/page-patterns/index.js
@@ -47,6 +47,9 @@ import {
 	OPERATOR_IS,
 } from '../../utils/constants';
 import {
+	exportJSONaction,
+	resetAction,
+	deleteAction,
 	duplicatePatternAction,
 	duplicateTemplatePartAction,
 } from './dataviews-pattern-actions';
@@ -380,13 +383,20 @@ export default function DataviewsPatterns() {
 		if ( type === TEMPLATE_PART_POST_TYPE ) {
 			return [
 				editAction,
-				duplicateTemplatePartAction,
 				...templatePartActions,
+				duplicateTemplatePartAction,
+				resetAction,
+				deleteAction,
 			].filter( Boolean );
 		}
-		return [ editAction, duplicatePatternAction, ...patternActions ].filter(
-			Boolean
-		);
+		return [
+			editAction,
+			...patternActions,
+			duplicatePatternAction,
+			exportJSONaction,
+			resetAction,
+			deleteAction,
+		].filter( Boolean );
 	}, [ editAction, type, templatePartActions, patternActions ] );
 	const onChangeView = useCallback(
 		( newView ) => {
