+++ packages/editor/package.json
@@ -64,8 +64,6 @@
 		"@wordpress/url": "file:../url",
 		"@wordpress/warning": "file:../warning",
 		"@wordpress/wordcount": "file:../wordcount",
-		"change-case": "^4.1.2",
-		"client-zip": "^2.4.4",
 		"clsx": "^2.1.1",
 		"date-fns": "^3.6.0",
 		"deepmerge": "^4.3.0",
